requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.14.0
google-generativeai==0.8.3
schedule==1.2.0
jinja2==3.1.3
email-validator==2.1.0
resend==2.4.0
markdown==3.5.1
pymdown-extensions==10.16